refactor: tags 상수들 constants로 분리
